pyproject.toml
[tool.poetry]
name = "dispander"
version = "0.6"
description = "Discord Message URL Expander"
authors = ["1ntegrale9 <1ntegrale9uation@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "~2.0.0"

[tool.poetry.dev-dependencies]
pyright = "^1.1.267"
